initial commit (just in case I need the code back)

+++ math/fancy_math.h
@@ -3,34 +3,35 @@
 
 #define _USE_MATH_DEFINES
 #include <math.h>
+#include <QtGlobal>
 #include "vector2.h"
 
 
 #define TOLERANCE 0.00001f
 
 
-inline float minf(float f1, float f2) {
+inline qreal minf(qreal f1, qreal f2) {
 	return f1 < f2 ? f1 : f2;
 }
 
 
-inline float maxf(float f1, float f2) {
+inline qreal maxf(qreal f1, qreal f2) {
 	return f1 > f2 ? f1 : f2;
 }
 
 
-inline float clampf(float f, float min, float max) {
+inline qreal clampf(qreal f, qreal min, qreal max) {
 	return f < min ? min : (f > max ? max : f);
 }
 
 
-inline bool is_zero_approx(float f) {
+inline bool is_zero_approx(qreal f) {
 	return abs(f) < TOLERANCE;
 }
 
 
-inline bool is_equal_approx(float f1, float f2) {
-	float x = maxf(f1, f2);
+inline bool is_equal_approx(qreal f1, qreal f2) {
+	qreal x = maxf(f1, f2);
 	return abs(f1 - x) < TOLERANCE;
 }
